requirements.txt
streamlit==1.31.0
pandas==2.0.3
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
openai==1.35.7
yfinance==0.2.40
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.93.0
markdown2==2.4.10
jinja2==3.1.2
toml==0.10.2